Fix item actions collapsing and delete dialog visibility, update item actions visibility
- Fix collapsing of other items actions not working correctly
- Collapse all items actions on navigation
- Update item actions visibility after delete dialog is shown or hidden

+++ app/src/main/java/com/luczka/mycoffee/ui/screens/coffeedetails/CoffeeDetailsScreen.kt
@@ -50,17 +50,19 @@ fun CoffeeDetailsScreen(
     onAction: (CoffeeDetailsAction) -> Unit
 ) {
     if (uiState is CoffeeDetailsUiState.HasCoffee) {
-        DeleteCoffeeDialog(
-            coffeeUiState = uiState.coffee,
-            onNegative = {
-                val action = CoffeeDetailsAction.HideDeleteDialog
-                onAction(action)
-            },
-            onPositive = {
-                val action = CoffeeDetailsAction.OnDeleteClicked
-                onAction(action)
-            }
-        )
+        if(uiState.openDeleteDialog) {
+            DeleteCoffeeDialog(
+                coffeeUiState = uiState.coffee,
+                onNegative = {
+                    val action = CoffeeDetailsAction.HideDeleteDialog
+                    onAction(action)
+                },
+                onPositive = {
+                    val action = CoffeeDetailsAction.OnDeleteClicked
+                    onAction(action)
+                }
+            )
+        }
     }
 
     Scaffold(
